+++ Packages/Sources/TileKit/TileKit.Markdown.Rendering.swift
@@ -1,0 +1,7 @@
+public extension TileKit.Markdown {
+    protocol Rendering {
+        func renderHTML(
+            _ markdown: String,
+        ) -> String
+    }
+}