+++ Packages/Sources/TileKit/TileKit.Source.FrontMatterParserError.swift
@@ -1,0 +1,6 @@
+public extension TileKit.Source {
+    enum FrontMatterParserError: Error, Equatable {
+        case missingClosingSeparator
+        case invalidLine(String)
+    }
+}